+++ src/Microsoft.Diagnostics.Runtime/ClrThread.cs
@@ -17,15 +17,20 @@ namespace Microsoft.Diagnostics.Runtime
     public enum ClrStackFrameType
     {
         /// <summary>
+        /// Indicates this stack frame is unknown
+        /// </summary>
+        Unknown = -1,
+
+        /// <summary>
         /// Indicates this stack frame is a standard managed method.
         /// </summary>
-        ManagedMethod,
+        ManagedMethod = 0,
 
         /// <summary>
         /// Indicates this stack frame is a special stack marker that the Clr runtime leaves on the stack.
         /// Note that the ClrStackFrame may still have a ClrMethod associated with the marker.
         /// </summary>
-        Runtime
+        Runtime = 1
     }
 
     /// <summary>
