+++ src/Microsoft.Diagnostics.Runtime/Desktop/runtimebase.cs
@@ -51,6 +51,11 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
             }
         }
 
+        public override IEnumerable<ClrException> EnumerateSerializedExceptions()
+        {
+            return new ClrException[0];
+        }
+
         public override IEnumerable<int> EnumerateGCThreads()
         {
             foreach (uint thread in _dataReader.EnumerateAllThreads())
