+++ src/Microsoft.Diagnostics.Runtime/Microsoft.Diagnostics.Runtime.csproj
@@ -117,6 +117,7 @@
     <Compile Include="ICorDebug\IMetadataImport.cs" />
     <Compile Include="internal.cs" />
     <Compile Include="ISymbolResolver.cs" />
+    <Compile Include="Native\NativeException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ClrException.cs" />
     <Compile Include="Native\NativeThread.cs" />
