+++ src/Microsoft.Diagnostics.Runtime/Native/NativeHeap.cs
@@ -40,7 +40,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             {
                 get
                 {
-                    return $"{_symbolName}";
+                    return _symbolName;
                 }
             }
 
@@ -64,7 +64,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             {
                 get
                 {
-                    throw new NotImplementedException("This information is unavailable when debugging Native code");
+                    return ClrStackFrameType.Unknown;
                 }
             }
 
@@ -72,7 +72,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             {
                 get
                 {
-                    throw new NotImplementedException("This information is unavailable when debugging Native code");
+                    return null;
                 }
             }
 
@@ -80,7 +80,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             {
                 get
                 {
-                    throw new NotImplementedException("The SP is not serialized, hence it is unavailable when debugging Native code");
+                    return 0;
                 }
             }
 
@@ -88,7 +88,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             {
                 get
                 {
-                    throw new NotImplementedException("The Thread is not serialized, hence it is unavailable when debugging Native code");
+                    return null;
                 }
             }
         }
@@ -125,14 +125,20 @@ namespace Microsoft.Diagnostics.Runtime.Native
 
         public override ClrType GetTypeByMethodTable(ulong methodTable, ulong componentMethodTable)
         {
+            if ((((int)methodTable) & 3) != 0)
+                methodTable &= ~3UL;
+
             if (componentMethodTable != 0)
                 return null;
 
+            ClrType clrType = null;
             int index;
             if (_indices.TryGetValue(methodTable, out index))
-                return _types[index];
+                clrType = _types[index];
+            else
+                clrType = ConstructObjectType(methodTable);
 
-            return null;
+            return clrType;
         }
 
         private NativeModule FindMrtModule()
@@ -154,121 +160,6 @@ namespace Microsoft.Diagnostics.Runtime.Native
             _types.Add(_free);
         }
 
-        public override ClrType GetClrTypeFromEE(ulong eeType)
-        {
-            if ((((int)eeType) & 3) != 0)
-                eeType &= ~3UL;
-
-            ClrType clrType = null;
-            int index;
-            if (_indices.TryGetValue(eeType, out index))
-                clrType = _types[index];
-            else
-                clrType = ConstructObjectType(eeType);
-
-            return clrType;
-        }
-
-        internal class NativeException : ClrException
-        {
-            private Address _ccwPtr;
-            private ClrType _type;
-            private ulong _hResult;
-            private ulong _threadId;
-            private ulong _exceptionId;
-            private ulong _innerExceptionId;
-            private ulong _nestingLevel;
-            private IList<ClrStackFrame> _stackFrames;
-
-            private ClrException _innerException;
-            
-            public NativeException(
-                            ClrType clrType,
-                            Address ccwPtr,
-                            ulong hResult,
-                            ulong threadId,
-                            ulong exceptionId,
-                            ulong innerExceptionId,
-                            ulong nestingLevel,
-                            IList<ClrStackFrame> stackFrames)
-            {
-                _type = clrType;
-                _ccwPtr = ccwPtr;
-                _hResult = hResult;
-                _threadId = threadId;
-                _exceptionId = exceptionId;
-                _innerExceptionId = innerExceptionId;
-                _nestingLevel = nestingLevel;
-                _stackFrames = stackFrames;
-            }
-
-            public override Address Address
-            {
-                get { return _ccwPtr; }
-            }
-            public override int HResult
-            {
-                get
-                {
-                    return (int)_hResult;
-                }
-            }
-
-            public override ClrType Type
-            {
-                get { return _type; }
-            }
-
-            public override string Message
-            {
-                get
-                {
-                    return "<Invalid Object>"; // we do not expect a Message in serialized exceptions, however, returning <Invalid Object> as the SOS does
-                }
-            }
-
-            public override ClrException Inner
-            {
-                get
-                {
-                    return _innerException;
-                }
-            }
-
-            internal void setInnerException(ClrException inner)
-            {
-                _innerException = inner;
-            }
-
-            public override IList<ClrStackFrame> StackTrace
-            {
-                get
-                {
-                    return _stackFrames;
-                }
-            }
-
-            public ulong InnerExceptionId
-            {
-                get { return _innerExceptionId; }
-            }
-
-            public ulong ThreadId
-            {
-                get { return _threadId; }
-            }
-
-            public ulong NestingLevel
-            {
-                get { return _nestingLevel; }
-            }
-
-            public ulong ExceptionId
-            {
-                get { return _exceptionId; }
-            }
-        }
-
         public override ClrType GetObjectType(ulong objRef)
         {
             ulong eeType;
@@ -283,7 +174,7 @@ namespace Microsoft.Diagnostics.Runtime.Native
             if (!cache.ReadPtr(objRef, out eeType))
                 return null;
 
-            ClrType last = this.GetClrTypeFromEE(eeType);
+            ClrType last = this.GetTypeByMethodTable(eeType);
             _lastObj = objRef;
             _lastType = last;
             return last;
